service unit tests
