+++ test/unit/service.spec.js
@@ -5,1497 +5,1762 @@ const Context = require("../../src/context");
 const ServiceBroker = require("../../src/service-broker");
 const { protectReject } = require("./utils");
 
-/**
- * TODO:
- * 	Should rewrite some tests because many test integration test instead of unit test.
- */
+describe("Test Service class", () => {
+	describe("Test constructor", () => {
 
-describe("Test Service constructor", () => {
+		const broker = new ServiceBroker({ logger: false });
 
-	const broker = new ServiceBroker({ logger: false });
+		jest.spyOn(Service.prototype, "parseServiceSchema");
 
-	const schema = {
-		name: "users",
-		version: 2,
-		settings: {
-			a: 1,
-			cache: true
-		}
-	};
+		it("should throw error if missing serviceBroker instance", () => {
+			expect(() => {
+				new Service();
+			}).toThrowError("Must set a ServiceBroker instance!");
+		});
 
-	broker.callMiddlewareHookSync = jest.fn();
-	jest.spyOn(broker, "getLogger");
-
-	it("should throw exceptions if missing main properties", () => {
-		expect(() => {
-			new Service();
-		}).toThrowError("Must set a ServiceBroker instance!");
-
-		expect(() => {
-			/* eslint-disable-next-line */
-			console.error = jest.fn();
+		it("should set broker to local variable", () => {
+			Service.prototype.parseServiceSchema.mockClear();
 			const svc = new Service(broker);
-			svc.parseServiceSchema();
-		}).toThrowError("Must pass a service schema in constructor. Maybe is it not a service schema?");
+			expect(svc.broker).toBe(broker);
+			expect(svc.Promise).toBe(broker.Promise);
+			expect(Service.prototype.parseServiceSchema).toBeCalledTimes(0);
+		});
 
-		expect(() => {
-			new Service(broker, {});
-		}).toThrowError("Service name can't be empty! Maybe it is not a valid Service schema.");
+		it("should call the parseServiceSchema", () => {
+			Service.prototype.parseServiceSchema.mockClear();
+			const schema = {
+				name: "posts"
+			};
+
+			const svc = new Service(broker, schema);
+			expect(svc.broker).toBe(broker);
+			expect(svc.Promise).toBe(broker.Promise);
+			expect(Service.prototype.parseServiceSchema).toBeCalledTimes(1);
+			expect(Service.prototype.parseServiceSchema).toBeCalledWith(schema);
+		});
+
 	});
 
-	it("check local properties", () => {
-		broker.callMiddlewareHookSync.mockClear();
-		broker.getLogger.mockClear();
+	describe("Test parseServiceSchema", () => {
 
-		const service = new Service(broker, schema);
+		const broker = new ServiceBroker({ logger: false });
 
-		expect(service.name).toBe("users");
-		expect(service.version).toBe(2);
-		expect(service.fullName).toBe("v2.users");
-		expect(service.settings).toBe(schema.settings);
-		expect(service.metadata).toEqual({});
-		expect(service.schema).toBe(schema);
-		expect(service.broker).toBe(broker);
-		expect(service.Promise).toBe(broker.Promise);
+		jest.spyOn(broker, "callMiddlewareHookSync");
+		jest.spyOn(broker, "getLogger");
+		jest.spyOn(Service, "applyMixins");
 
-		expect(service.logger).toBeDefined();
-		expect(service.actions).toEqual({});
+		const svc = new Service(broker);
 
-		expect(broker.getLogger).toHaveBeenCalledTimes(1);
-		expect(broker.getLogger).toHaveBeenCalledWith("v2.users", {
-			svc: "users",
-			ver: 2
+		jest.spyOn(svc, "_init");
+
+		it("should throw error if missing schema", () => {
+			expect(() => {
+				svc.parseServiceSchema();
+			}).toThrowError("The service schema can't be null. Maybe is it not a service schema?");
 		});
 
-		expect(broker.callMiddlewareHookSync).toHaveBeenCalledTimes(2);
-		expect(broker.callMiddlewareHookSync).toHaveBeenNthCalledWith(1, "serviceCreating", [service, schema]);
-		expect(broker.callMiddlewareHookSync).toHaveBeenNthCalledWith(2, "serviceCreated", [service]);
-	});
-
-	it("check local properties with metadata", () => {
-		schema.metadata = {
-			scalable: true
-		};
-
-		const service = new Service(broker, schema);
-
-		expect(service.name).toBe("users");
-		expect(service.version).toBe(2);
-		expect(service.settings).toBe(schema.settings);
-		expect(service.metadata).toEqual({ scalable: true });
-		expect(service.schema).toBe(schema);
-		expect(service.broker).toBe(broker);
-
-		expect(service.logger).toBeDefined();
-		expect(service.actions).toEqual({});
-
-		expect(service.originalSchema).toEqual(schema);
-		expect(service.originalSchema).not.toBe(schema);
-	});
-
-	it("check empty properties", () => {
-		const service = new Service(broker, {
-			name: "empty",
-			mixins: [],
-			dependencies: []
+		it("should throw error if name is empty", () => {
+			Service.applyMixins.mockClear();
+			expect(() => {
+				/* eslint-disable-next-line */
+				console.error = jest.fn();
+				svc.parseServiceSchema({});
+			}).toThrowError("Service name can't be empty! Maybe it is not a valid Service schema. Maybe is it not a service schema?");
+			expect(Service.applyMixins).toBeCalledTimes(0);
 		});
 
-		expect(service.name).toBe("empty");
-		expect(service.broker).toBe(broker);
-	});
+		it("should set common local variables", () => {
+			Service.applyMixins.mockClear();
 
-	it("check logger with string version", () => {
-		broker.getLogger.mockClear();
-		const service = new Service(broker, {
-			name: "posts",
-			version: "stage"
+			const schema = { name: "posts" };
+			svc.parseServiceSchema(schema);
+
+			expect(svc.originalSchema).toEqual({ name: "posts" });
+			expect(svc.originalSchema).not.toBe(schema);
+
+			expect(Service.applyMixins).toBeCalledTimes(0);
+
+			expect(svc.name).toBe("posts");
+			expect(svc.version).toBeUndefined();
+			expect(svc.settings).toEqual({});
+			expect(svc.metadata).toEqual({});
+			expect(svc.schema).toBe(schema);
+			expect(svc.fullName).toBe("posts");
+			expect(svc.logger).toBeDefined();
+
+			expect(broker.getLogger).toBeCalledTimes(1);
+			expect(broker.getLogger).toBeCalledWith("posts", { svc: "posts", ver: undefined });
+
+			expect(svc.actions).toEqual({});
+			expect(svc.events).toEqual({});
+
+			expect(svc._init).toBeCalledTimes(1);
 		});
-		expect(service.fullName).toBe("stage.posts");
-		expect(broker.getLogger).toHaveBeenCalledTimes(1);
-		expect(broker.getLogger).toHaveBeenCalledWith("stage.posts", {
-			svc: "posts",
-			ver: "stage"
+
+		it("should set common local variables with version", () => {
+			Service.applyMixins.mockClear();
+			broker.getLogger.mockClear();
+			svc._init.mockClear();
+
+			const schema = { name: "posts", version: 3, mixins: [] };
+			svc.parseServiceSchema(schema);
+
+			expect(svc.originalSchema).toEqual({ name: "posts", version: 3, mixins: [] });
+			expect(svc.originalSchema).not.toBe(schema);
+
+			expect(Service.applyMixins).toBeCalledTimes(1);
+			expect(Service.applyMixins).toBeCalledWith(schema);
+
+			expect(svc.name).toBe("posts");
+			expect(svc.version).toBe(3);
+			expect(svc.settings).toEqual({});
+			expect(svc.metadata).toEqual({});
+			expect(svc.schema).toBe(schema);
+			expect(svc.fullName).toBe("v3.posts");
+			expect(svc.logger).toBeDefined();
+
+			expect(broker.getLogger).toBeCalledTimes(1);
+			expect(broker.getLogger).toBeCalledWith("v3.posts", { svc: "posts", ver: 3 });
+
+			expect(svc.actions).toEqual({});
+			expect(svc.events).toEqual({});
+
+			expect(svc._init).toBeCalledTimes(1);
 		});
-	});
 
-	it("check logger with string version but disabled versionPrefix", () => {
-		broker.getLogger.mockClear();
-		const service = new Service(broker, {
-			name: "posts",
-			version: "stage",
-			settings: {
-				$noVersionPrefix: true
-			}
+		it("should set common local variables with version & noVersionPrefix", () => {
+			Service.applyMixins.mockClear();
+			broker.getLogger.mockClear();
+			svc._init.mockClear();
+
+			const schema = { name: "posts", version: 3, settings: { $noVersionPrefix: true } };
+			svc.parseServiceSchema(schema);
+
+			expect(svc.name).toBe("posts");
+			expect(svc.version).toBe(3);
+			expect(svc.fullName).toBe("posts");
 		});
-		expect(service.fullName).toBe("posts");
-		expect(broker.getLogger).toHaveBeenCalledTimes(1);
-		expect(broker.getLogger).toHaveBeenCalledWith("posts", {
-			svc: "posts",
-			ver: "stage"
-		});
-	});
 
-	it("check logger without version", () => {
-		broker.getLogger.mockClear();
-		const service = new Service(broker, {
-			name: "likes"
-		});
-		expect(service.fullName).toBe("likes");
-		expect(broker.getLogger).toHaveBeenCalledTimes(1);
-		expect(broker.getLogger).toHaveBeenCalledWith("likes", {
-			svc: "likes",
-			ver: undefined
-		});
-	});
-
-});
-
-describe("Test action creation", () => {
-	const broker = new ServiceBroker({ logger: false, internalServices: false });
-	broker.callMiddlewareHook = jest.fn();
-
-	const schema = {
-		name: "posts",
-		actions: {
-			find: jest.fn(),
-			get: {
-				cache: false,
-				params: {
-					id: "number"
-				},
-				handler: jest.fn()
-			},
-			noExist: false
-		}
-	};
-
-	it("should register service & actions", () => {
-		broker.addLocalService = jest.fn();
-
-		const ep = {};
-		broker.registry.createPrivateActionEndpoint = jest.fn(() => ep);
-
-		let service = broker.createService(schema);
-
-		expect(broker.registry.createPrivateActionEndpoint).toHaveBeenCalledTimes(2);
-
-		expect(service).toBeDefined();
-
-		expect(broker.addLocalService).toHaveBeenCalledTimes(1);
-		expect(broker.addLocalService.mock.calls[0][0]).toBe(service);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => {
-				const spec = service._serviceSpecification;
-				expect(spec).toEqual({
+		it("should throw error if method name is now allowed", () => {
+			expect(() => {
+				svc.parseServiceSchema({
 					name: "posts",
-					version: undefined,
-					fullName: "posts",
-					settings: {},
-					metadata: {},
-					actions: {
-						"posts.find": jasmine.any(Object),
-						"posts.get": jasmine.any(Object)
-					},
-					events: {}
+					methods: {
+						actions() {}
+					}
 				});
-
-				expect(service.actions.find).toBeDefined();
-				expect(service.actions.get).toBeDefined();
-
-				let ctx = new Context(broker);
-				let oldCreate = broker.ContextFactory.create;
-				broker.ContextFactory.create = jest.fn(() => ctx);
-
-				service.actions.find({ a: 5 }, { timeout: 1000 });
-
-				expect(broker.ContextFactory.create).toHaveBeenCalledTimes(1);
-				expect(broker.ContextFactory.create).toHaveBeenCalledWith(broker, ep, { a: 5 }, { timeout: 1000 });
-
-				expect(schema.actions.find).toHaveBeenCalledTimes(1);
-				expect(schema.actions.find).toHaveBeenCalledWith(ctx);
-
-				// ---
-				broker.ContextFactory.create.mockClear();
-				let prevCtx = new Context(broker, { action: service.schema.actions.get });
-				prevCtx.setParams({ id: 10 });
-				service.actions.get({ id: 20 }, { ctx: prevCtx });
-
-				expect(broker.ContextFactory.create).toHaveBeenCalledTimes(0);
-
-				expect(schema.actions.get.handler).toHaveBeenCalledTimes(1);
-				expect(schema.actions.get.handler).toHaveBeenCalledWith(prevCtx);
-
-				broker.ContextFactory.create = oldCreate;
-			});
-
-	});
-
-	it("should throw error if action is neither object nor function", () => {
-		expect(() => {
-			broker.createService({
-				name: "test",
-				actions: {
-					hello: 500
-				}
-			});
-		}).toThrowError("Invalid action definition in 'hello' action in 'test' service!");
-	});
-
-});
-
-describe("Test service start", () => {
-
-	it("should start service with simple handler", () => {
-		let broker = new ServiceBroker({ logger: false, internalServices: false });
-		broker.callMiddlewareHook = jest.fn();
-		broker.registerLocalService = jest.fn();
-
-		let schema = {
-			name: "posts",
-			started: jest.fn()
-		};
-
-		let service = broker.createService(schema);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => {
-				expect(schema.started).toHaveBeenCalledTimes(1);
-
-				expect(broker.callMiddlewareHook).toHaveBeenCalledTimes(2);
-				expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(1, "serviceStarting", [service]);
-				expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(2, "serviceStarted", [service]);
-
-				expect(broker.registerLocalService).toHaveBeenCalledTimes(1);
-				expect(broker.registerLocalService.mock.calls[0][0]).toBe(service._serviceSpecification);
-
-			});
-	});
-
-	it("should start service with multiple handler", () => {
-		let broker = new ServiceBroker({ logger: false, internalServices: false });
-		broker.callMiddlewareHook = jest.fn();
-		broker.registerLocalService = jest.fn();
-
-		let schema = {
-			name: "posts",
-			started: [jest.fn(), jest.fn()]
-		};
-
-		let service = broker.createService(schema);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => {
-				expect(schema.started[0]).toHaveBeenCalledTimes(1);
-				expect(schema.started[1]).toHaveBeenCalledTimes(1);
-
-				expect(broker.callMiddlewareHook).toHaveBeenCalledTimes(2);
-				expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(1, "serviceStarting", [service]);
-				expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(2, "serviceStarted", [service]);
-
-				expect(broker.registerLocalService).toHaveBeenCalledTimes(1);
-				expect(broker.registerLocalService.mock.calls[0][0]).toBe(service._serviceSpecification);
-
-			});
-
-	});
-});
-
-describe("Test service stop", () => {
-
-	it("should stop service with simple handler", () => {
-		let broker = new ServiceBroker({ logger: false, internalServices: false });
-		broker.callMiddlewareHook = jest.fn();
-
-		let schema = {
-			name: "posts",
-			stopped: jest.fn()
-		};
-
-		let service = broker.createService(schema);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => {
-				broker.callMiddlewareHook.mockClear();
-
-				return service._stop()
-					.catch(protectReject)
-					.then(() => {
-						expect(schema.stopped).toHaveBeenCalledTimes(1);
-
-						expect(broker.callMiddlewareHook).toHaveBeenCalledTimes(2);
-						expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(1, "serviceStopping", [service], { reverse: true });
-						expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(2, "serviceStopped", [service], { reverse: true });
-
-					});
-			});
-	});
-
-	it("should stop service with multiple handler", () => {
-		let broker = new ServiceBroker({ logger: false, internalServices: false });
-		broker.callMiddlewareHook = jest.fn();
-
-		let schema = {
-			name: "posts",
-			stopped: [jest.fn(), jest.fn()]
-		};
-
-		let service = broker.createService(schema);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => {
-				broker.callMiddlewareHook.mockClear();
-
-				return service._stop()
-					.catch(protectReject)
-					.then(() => {
-						expect(schema.stopped[0]).toHaveBeenCalledTimes(1);
-						expect(schema.stopped[1]).toHaveBeenCalledTimes(1);
-
-						expect(broker.callMiddlewareHook).toHaveBeenCalledTimes(2);
-						expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(1, "serviceStopping", [service], { reverse: true });
-						expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(2, "serviceStopped", [service], { reverse: true });
-
-					});
-			});
-
-	});
-});
-
-describe("Test events creation", () => {
-	it("should register event handler to broker", () => {
-		const broker = new ServiceBroker({ logger: false, internalServices: false });
-		broker.registerLocalService = jest.fn();
-		broker.addLocalService = jest.fn();
-
-		const service = broker.createService({
-			name: "posts",
-			events: {
-				"user.*": jest.fn(),
-				"posts.updated": {
-					handler: jest.fn()
-				}
-			}
+			}).toThrowError("Invalid method name 'actions' in 'posts' service!");
 		});
 
-		expect(service).toBeDefined();
-
-		expect(broker.addLocalService).toHaveBeenCalledTimes(1);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => {
-				expect(broker.registerLocalService).toHaveBeenCalledTimes(1);
-				expect(broker.registerLocalService.mock.calls[0][0]).toBe(service._serviceSpecification);
-
-				const spec = service._serviceSpecification;
-				expect(spec.events["posts.updated"]).toBeDefined();
-				expect(spec.events["user.*"]).toBeDefined();
-				expect(spec.actions).toEqual({});
-			});
-	});
-
-	it("should register event handler with mixins", () => {
-		let broker = new ServiceBroker({ logger: false, internalServices: false });
-		broker.addLocalService = jest.fn();
-		broker.registerLocalService = jest.fn();
-
-		const cb1 = jest.fn();
-		const cb2 = jest.fn();
-		const cb3 = jest.fn();
-		const service = broker.createService({
-			name: "posts",
-			events: {
-				"user.*": [cb1, ctx => cb2(ctx)],
-				"posts.updated": {
-					handler: cb3
-				}
-			}
-		});
-
-		expect(service).toBeDefined();
-
-		expect(broker.addLocalService).toHaveBeenCalledTimes(1);
-		return service._start()
-			.catch(protectReject)
-			.then(() => {
-				expect(broker.registerLocalService).toHaveBeenCalledTimes(1);
-				expect(broker.registerLocalService.mock.calls[0][0]).toBe(service._serviceSpecification);
-
-				const spec = service._serviceSpecification;
-				expect(spec.events["posts.updated"]).toBeDefined();
-				expect(spec.events["user.*"]).toBeDefined();
-				expect(spec.actions).toEqual({});
-
-				const ctx = { params: { a: 5 }, nodeID: "node-123", eventName: "posts.updated" };
-				spec.events["posts.updated"].handler(ctx);
-				expect(cb3).toHaveBeenCalledTimes(1);
-				expect(cb3).toHaveBeenCalledWith({ a: 5 }, "node-123", "posts.updated", ctx);
-
-				spec.events["user.*"].handler(ctx);
-				expect(cb1).toHaveBeenCalledTimes(1);
-				expect(cb1).toHaveBeenCalledWith({ a: 5 }, "node-123", "posts.updated", ctx);
-
-				expect(cb2).toHaveBeenCalledTimes(1);
-				expect(cb2).toHaveBeenCalledWith(ctx);
-			});
-	});
-
-	it("should throw error because no handler of event", () => {
-		let broker = new ServiceBroker({ logger: false, internalServices: false });
-		expect(() => {
-			broker.createService({
-				name: "test",
-				events: {
-					"register.node": {}
-				}
-			});
-		}).toThrowError("Missing event handler on 'register.node' event in 'test' service!");
-	});
-
-	it("should call event handlers directly", () => {
-		const broker = new ServiceBroker({ logger: false, internalServices: false });
-		broker.registerLocalService = jest.fn();
-		broker.addLocalService = jest.fn();
-
-		const userEventHandler = jest.fn();
-		const postEventHandler = jest.fn(() => Promise.resolve());
-		const service = broker.createService({
-			name: "posts",
-			events: {
-				"user.*": userEventHandler,
-				"posts.updated": {
-					context: true,
-					handler: postEventHandler
-				}
-			}
-		});
-
-		expect(service).toBeDefined();
-		expect(service.events).toEqual({
-			"user.*": expect.any(Function),
-			"posts.updated": expect.any(Function),
-		});
-
-		expect(broker.addLocalService).toHaveBeenCalledTimes(1);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => service.events["user.*"]({ a: 5 }))
-			.catch(protectReject)
-			.then(() => {
-				expect(userEventHandler).toHaveBeenCalledTimes(1);
-				expect(userEventHandler).toHaveBeenCalledWith({ a: 5 }, broker.nodeID, "user.*", expect.any(Context));
-				expect(userEventHandler.mock.calls[0][3].params).toEqual({ a: 5 });
-				expect(userEventHandler.mock.calls[0][3].eventName).toEqual("user.*");
-				expect(userEventHandler.mock.calls[0][3].eventGroups).toEqual(["posts"]);
-				expect(userEventHandler.mock.calls[0][3].eventType).toEqual("emit");
-			})
-
-			.then(() => service.events["posts.updated"]({ b: "John" }, { requestID: "12345" }))
-			.then(() => {
-				expect(postEventHandler).toHaveBeenCalledTimes(1);
-				expect(postEventHandler).toHaveBeenCalledWith(expect.any(Context));
-				expect(postEventHandler.mock.calls[0][0].params).toEqual({ b: "John" });
-				expect(postEventHandler.mock.calls[0][0].eventName).toEqual("posts.updated");
-				expect(postEventHandler.mock.calls[0][0].eventGroups).toEqual(["posts"]);
-				expect(postEventHandler.mock.calls[0][0].eventType).toEqual("emit");
-				expect(postEventHandler.mock.calls[0][0].requestID).toEqual("12345");
-			});
-	});
-
-	it("should call event handlers directly with emitLocalEventHandler", () => {
-		const broker = new ServiceBroker({ logger: false, internalServices: false });
-		broker.registerLocalService = jest.fn();
-		broker.addLocalService = jest.fn();
-
-		const userEventHandler = jest.fn();
-		const postEventHandler = jest.fn(() => Promise.resolve());
-		const service = broker.createService({
-			name: "posts",
-			events: {
-				"user.*": userEventHandler,
-				"posts.updated": {
-					context: true,
-					handler: postEventHandler
-				}
-			}
-		});
-
-		expect(service).toBeDefined();
-		expect(service.events).toEqual({
-			"user.*": expect.any(Function),
-			"posts.updated": expect.any(Function),
-		});
-
-		expect(broker.addLocalService).toHaveBeenCalledTimes(1);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => service.emitLocalEventHandler("user.*", { a: 5 }))
-			.catch(protectReject)
-			.then(() => {
-				expect(userEventHandler).toHaveBeenCalledTimes(1);
-				expect(userEventHandler).toHaveBeenCalledWith({ a: 5 }, broker.nodeID, "user.*", expect.any(Context));
-				expect(userEventHandler.mock.calls[0][3].params).toEqual({ a: 5 });
-				expect(userEventHandler.mock.calls[0][3].eventName).toEqual("user.*");
-				expect(userEventHandler.mock.calls[0][3].eventGroups).toEqual(["posts"]);
-				expect(userEventHandler.mock.calls[0][3].eventType).toEqual("emit");
-			})
-
-			.then(() => service.emitLocalEventHandler("posts.updated", { b: "John" }, { requestID: "12345" }))
-			.then(() => {
-				expect(postEventHandler).toHaveBeenCalledTimes(1);
-				expect(postEventHandler).toHaveBeenCalledWith(expect.any(Context));
-				expect(postEventHandler.mock.calls[0][0].params).toEqual({ b: "John" });
-				expect(postEventHandler.mock.calls[0][0].eventName).toEqual("posts.updated");
-				expect(postEventHandler.mock.calls[0][0].eventGroups).toEqual(["posts"]);
-				expect(postEventHandler.mock.calls[0][0].eventType).toEqual("emit");
-				expect(postEventHandler.mock.calls[0][0].requestID).toEqual("12345");
-			});
-	});
-});
-
-describe("Test methods creation", () => {
-	it("should create method in Service instance", () => {
-		let broker = new ServiceBroker({ logger: false, internalServices: false });
-		let service = broker.createService({
-			name: "posts",
-			methods: {
-				something: jest.fn()
-			}
-		});
-
-		expect(service).toBeDefined();
-		expect(typeof service.something).toBe("function");
-
-		service.something();
-
-		expect(service.schema.methods.something).toHaveBeenCalledTimes(1);
-	});
-
-	it("should throw error because method name is reserved", () => {
-		let broker = new ServiceBroker({ logger: false, internalServices: false });
-		expect(() => {
-			broker.createService({
-				name: "test",
+		it("should set methods and bind the service instance", () => {
+			const schema = {
+				name: "posts",
 				methods: {
-					name: jest.fn()
+					alpha: () => "Alpha",
+					beta() {
+						expect(this).toBe(svc);
+						return "Beta";
+					}
+				}
+			};
+			svc.parseServiceSchema(schema);
+
+			expect(svc.alpha).toBeInstanceOf(Function);
+			expect(svc.beta).toBeInstanceOf(Function);
+
+			expect(svc.alpha()).toBe("Alpha");
+			expect(svc.beta()).toBe("Beta");
+
+			expect(svc._serviceSpecification).toEqual({
+				name: "posts",
+				version: undefined,
+				fullName: "posts",
+				settings: {},
+				metadata: {},
+				actions: {},
+				events: {}
+			});
+		});
+
+		it("should register actions", () => {
+			jest.spyOn(svc, "_createAction").mockImplementation((action, name) => ({ name, handler: action.handler || action }));
+			jest.spyOn(broker.middlewares, "wrapHandler").mockImplementation((type, handler, action) => handler);
+			jest.spyOn(broker.registry, "createPrivateActionEndpoint").mockImplementation(() => ({ id: "nodeID" }));
+			const schema = {
+				name: "posts",
+				actions: {
+					alpha: jest.fn(() => "Alpha"),
+					beta: false,
+					gamma: {
+						handler: jest.fn()
+					}
+				}
+			};
+			svc.parseServiceSchema(schema);
+
+			expect(svc._createAction).toBeCalledTimes(2);
+			expect(svc._createAction).toHaveBeenNthCalledWith(1, schema.actions.alpha, "alpha");
+			expect(svc._createAction).toHaveBeenNthCalledWith(2, schema.actions.gamma, "gamma");
+
+			expect(broker.middlewares.wrapHandler).toBeCalledTimes(2);
+			expect(broker.middlewares.wrapHandler).toHaveBeenNthCalledWith(1, "localAction", expect.any(Function), { name: "alpha", handler: expect.any(Function) });
+			expect(broker.middlewares.wrapHandler).toHaveBeenNthCalledWith(2, "localAction", expect.any(Function), { name: "gamma", handler: expect.any(Function) });
+
+			expect(broker.registry.createPrivateActionEndpoint).toBeCalledTimes(2);
+			expect(broker.registry.createPrivateActionEndpoint).toHaveBeenNthCalledWith(1, { name: "alpha", handler: expect.any(Function) });
+			expect(broker.registry.createPrivateActionEndpoint).toHaveBeenNthCalledWith(2, { name: "gamma", handler: expect.any(Function) });
+
+			expect(svc._serviceSpecification).toEqual({
+				name: "posts",
+				version: undefined,
+				fullName: "posts",
+				settings: {},
+				metadata: {},
+				actions: {
+					alpha: {
+						name: "alpha",
+						handler: expect.any(Function)
+					},
+					gamma: {
+						name: "gamma",
+						handler: expect.any(Function)
+					}
+				},
+				events: {}
+			});
+
+			expect(svc.actions.alpha).toBeInstanceOf(Function);
+			expect(svc.actions.beta).toBeUndefined();
+			expect(svc.actions.gamma).toBeInstanceOf(Function);
+		});
+
+		it("should call the handler via this.actions.xy", () => {
+			const schema = {
+				name: "posts",
+				actions: {
+					alpha: jest.fn(() => "Alpha"),
+				}
+			};
+			svc.parseServiceSchema(schema);
+
+			const res = svc.actions.alpha({ a: 5 });
+
+			expect(res).toBe("Alpha");
+			expect(schema.actions.alpha).toBeCalledTimes(1);
+			expect(schema.actions.alpha).toBeCalledWith(expect.any(Context));
+			expect(schema.actions.alpha.mock.calls[0][0].params).toEqual({ a: 5 });
+			expect(schema.actions.alpha.mock.calls[0][0].endpoint).toEqual({ id: "nodeID" });
+
+			schema.actions.alpha.mockClear();
+			const fakeCtx = { id: 123 };
+			const res2 = svc.actions.alpha({ a: 5 }, { ctx: fakeCtx });
+
+			expect(res2).toBe("Alpha");
+			expect(schema.actions.alpha).toBeCalledTimes(1);
+			expect(schema.actions.alpha).toBeCalledWith(fakeCtx);
+		});
+
+		it("should register events", () => {
+			jest.spyOn(svc, "_createEvent").mockImplementation((event, name) => ({ name, handler: event.handler || event }));
+
+			const schema = {
+				name: "posts",
+				events: {
+					"user.created": jest.fn(() => "User created"),
+					"posts.updated": {
+						handler: jest.fn(() => "Post #1 updated")
+					}
+				}
+			};
+			svc.parseServiceSchema(schema);
+
+			expect(svc._createEvent).toBeCalledTimes(2);
+			expect(svc._createEvent).toHaveBeenNthCalledWith(1, schema.events["user.created"], "user.created");
+			expect(svc._createEvent).toHaveBeenNthCalledWith(2, schema.events["posts.updated"], "posts.updated");
+
+			expect(svc._serviceSpecification).toEqual({
+				name: "posts",
+				version: undefined,
+				fullName: "posts",
+				settings: {},
+				metadata: {},
+				actions: {},
+				events: {
+					"user.created": {
+						name: "user.created",
+						handler: expect.any(Function)
+					},
+					"posts.updated": {
+						name: "posts.updated",
+						handler: expect.any(Function)
+					}
 				}
 			});
-		}).toThrowError("Invalid method name 'name' in 'test' service!");
-	});
-});
 
-describe("Test created event handler", () => {
-	let broker = new ServiceBroker({ logger: false, internalServices: false });
-
-	let schema = {
-		name: "posts",
-		created: jest.fn()
-	};
-
-	it("should create method in Service instance", () => {
-		broker.createService(schema);
-		expect(schema.created).toHaveBeenCalledTimes(1);
-	});
-});
-
-describe("Test _createAction function", () => {
-	let broker = new ServiceBroker({ logger: false });
-
-	const handler = jest.fn();
-
-	it("should create action object with default values", () => {
-		let service = broker.createService({ name: "users" });
-
-		let action = service._createAction({ handler }, "find");
-		expect(action.name).toBe("users.find");
-		expect(action.rawName).toBe("find");
-		expect(action.cache).toBe(false);
-		expect(action.handler).toBeInstanceOf(Function);
-		expect(action.service).toBe(service);
-	});
-
-	it("should create action with version number", () => {
-		let service = broker.createService({ name: "users", version: 3 });
-
-		let action = service._createAction({ handler, myProp: "teszt" }, "find");
-		expect(action.name).toBe("v3.users.find");
-		expect(action.rawName).toBe("find");
-		expect(action.myProp).toBe("teszt");
-	});
-
-	it("should create action with version 0", () => {
-		let service = broker.createService({ name: "users", version: 0 });
-
-		let action = service._createAction({ handler }, "find");
-		expect(action.name).toBe("v0.users.find");
-		expect(action.rawName).toBe("find");
-	});
-
-	it("should create action with version string", () => {
-		let service = broker.createService({ name: "users", version: "staging" });
-
-		let action = service._createAction({ handler }, "find");
-		expect(action.name).toBe("staging.users.find");
-		expect(action.rawName).toBe("find");
-	});
-
-	it("should create action without version", () => {
-		let service = broker.createService({ name: "users", version: 2, settings: { $noVersionPrefix: true } });
-
-		let action = service._createAction({ handler }, "find");
-		expect(action.name).toBe("users.find");
-		expect(action.rawName).toBe("find");
-	});
-
-	it("should create action with different name", () => {
-		let service = broker.createService({ name: "users" });
-
-		let action = service._createAction({ handler, name: "list" }, "find");
-		expect(action.name).toBe("users.list");
-		expect(action.rawName).toBe("list");
-	});
-
-	it("should create action without service name", () => {
-		let service = broker.createService({
-			name: "users",
-			settings: {
-				$noServiceNamePrefix: true
-			}
+			expect(svc.events["user.created"]).toBeInstanceOf(Function);
+			expect(svc.events["posts.updated"]).toBeInstanceOf(Function);
 		});
 
-		let action = service._createAction({ handler }, "find");
-		expect(action.name).toBe("find");
-		expect(action.rawName).toBe("find");
-	});
-
-
-	it("should throw Error if no handler", () => {
-		let service = broker.createService({ name: "users" });
-
-		expect(() => service._createAction({}, "find")).toThrowError("Missing action handler on 'find' action in 'users' service!");
-	});
-
-	describe("Test action cache property", () => {
-
-		it("cache is FALSE, if schema cache is UNDEFINED and action cache is UNDEFINED", () => {
-			let service = broker.createService({ name: "test" });
-			let action = service._createAction({ handler }, "find");
-
-			expect(action.cache).toBe(false);
-		});
-
-		it("cache is TRUE, if schema cache is TRUE and action cache UNDEFINED", () => {
-			let service = broker.createService({ name: "test", settings: { $cache: true } });
-			let action = service._createAction({ handler }, "find");
-
-			expect(action.cache).toBe(true);
-		});
-
-		it("cache is FALSE, if schema cache is TRUE and action cache is FALSE", () => {
-			let service = broker.createService({ name: "test", settings: { $cache: true } });
-			let action = service._createAction({ handler, cache: false }, "find");
-
-			expect(action.cache).toBe(false);
-		});
-
-		it("cache is TRUE, if schema cache is UNDEFINED and action cache is TRUE", () => {
-			let service = broker.createService({ name: "test" });
-			let action = service._createAction({ handler, cache: true }, "find");
-
-			expect(action.cache).toBe(true);
-		});
-
-		it("cache is TRUE, if schema cache is UNDEFINED and action cache is Object", () => {
-			let service = broker.createService({ name: "test" });
-			let action = service._createAction({ handler, cache: {} }, "find");
-
-			expect(action.cache).toEqual({});
-		});
-
-		it("cache is TRUE, if schema cache is FALSE and action cache is TRUE", () => {
-			let service = broker.createService({ name: "test", settings: { $cache: false } });
-			let action = service._createAction({ handler, cache: true }, "find");
-
-			expect(action.cache).toBe(true);
-		});
-
-		it("cache is TRUE, if schema cache is TRUE and action cache is TRUE", () => {
-			let service = broker.createService({ name: "test", settings: { $cache: true } });
-			let action = service._createAction({ handler, cache: true }, "find");
-
-			expect(action.cache).toBe(true);
-		});
-
-		it("cache is TRUE, if schema cache is TRUE and action cache is Object", () => {
-			let service = broker.createService({ name: "test", settings: { $cache: true } });
-			let action = service._createAction({ handler, cache: { keys: [ "id" ] } }, "find");
-
-			expect(action.cache).toEqual({ keys: [ "id" ] });
-		});
-
-	});
-});
-
-
-describe("Test constructor with mixins", () => {
-	const broker = new ServiceBroker({ logger: false });
-
-	const mixin1 = { name: "mixin1" };
-	const mixin2 = { name: "mixin2" };
-
-	const schema = {
-		name: "posts",
-		mixins: [
-			mixin1,
-			mixin2
-		]
-	};
-
-	it("should call applyMixins with schema", () => {
-		let oldApply = Service.applyMixins;
-		Service.applyMixins = jest.fn(schema => schema);
-
-		const svc = new Service(broker, schema);
-
-		expect(Service.applyMixins).toHaveBeenCalledTimes(1);
-		expect(Service.applyMixins).toHaveBeenCalledWith(schema);
-
-		Service.applyMixins = oldApply;
-	});
-});
-
-describe("Test lifecycle event handlers", () => {
-	describe("with simple handlers", () => {
-
-		let broker = new ServiceBroker({ logger: false });
-
-		let schema = {
-			name: "simple",
-
-			created: jest.fn(),
-			started: jest.fn(),
-			stopped: jest.fn()
-		};
-
-		it("should called created", () => {
-			broker.createService(schema);
-
-			expect(schema.created).toHaveBeenCalledTimes(1);
-
-			return broker.start();
-		});
-
-		it("should called started", () => {
-			expect(schema.started).toHaveBeenCalledTimes(1);
-
-			return broker.stop();
-		});
-
-		it("should called stopped", () => {
-			expect(schema.stopped).toHaveBeenCalledTimes(1);
-		});
-	});
-
-	describe("with multiple handlers (from mixins)", () => {
-
-		let broker = new ServiceBroker({ logger: false });
-
-		let FLOW = [];
-
-		let createdFn1 = jest.fn(() => FLOW.push("C1"));
-		let createdFn2 = jest.fn(() => FLOW.push("C2"));
-
-		let startedFn1 = jest.fn(() => FLOW.push("A1"));
-		let startedFn2 = jest.fn(() => FLOW.push("A2"));
-
-		let stoppedFn1 = jest.fn(() => FLOW.push("O1"));
-		let stoppedFn2 = jest.fn(() => FLOW.push("O2"));
-
-		let schema = {
-			name: "simple",
-
-			created: [
-				createdFn1,
-				createdFn2
-			],
-			started: [
-				startedFn1,
-				startedFn2
-			],
-			stopped: [
-				stoppedFn1,
-				stoppedFn2
-			],
-		};
-
-		it("should called created", () => {
-			broker.createService(schema);
-
-			expect(createdFn1).toHaveBeenCalledTimes(1);
-			expect(createdFn2).toHaveBeenCalledTimes(1);
-
-			return broker.start();
-		});
-
-		it("should called started", () => {
-			expect(startedFn1).toHaveBeenCalledTimes(1);
-			expect(startedFn2).toHaveBeenCalledTimes(1);
-
-			return broker.stop();
-		});
-
-		it("should called stopped", () => {
-			expect(stoppedFn1).toHaveBeenCalledTimes(1);
-			expect(stoppedFn2).toHaveBeenCalledTimes(1);
-
-			expect(FLOW.join("-")).toBe("C1-C2-A1-A2-O2-O1");
-		});
-	});
-
-});
-
-describe("Test broker.waitForServices", () => {
-	let broker = new ServiceBroker({ logger: false });
-	broker.waitForServices = jest.fn();
-
-	it("should call waitForServices", () => {
-		let svc = broker.createService({
-			name: "test",
-			created() {
-				this.waitForServices("posts", 5000, 500);
-			}
-		});
-		expect(broker.waitForServices).toHaveBeenCalledTimes(1);
-		expect(broker.waitForServices).toHaveBeenCalledWith("posts", 5000, 500, svc.logger);
-	});
-
-});
-
-describe("Test dependencies", () => {
-
-	describe("with one dependency", () => {
-
-		let broker = new ServiceBroker({ logger: false });
-
-		let schema = {
-			name: "simple",
-			dependencies: "math",
-
-			created: jest.fn(),
-			started: jest.fn()
-		};
-		let svc;
-
-		it("should called created", () => {
-			svc = broker.createService(schema);
-			svc.waitForServices = jest.fn(() => Promise.resolve());
-
-			expect(schema.created).toHaveBeenCalledTimes(1);
-
-			return broker.start();
-		});
-
-		it("should called started", () => {
-			expect(schema.started).toHaveBeenCalledTimes(1);
-			expect(svc.waitForServices).toHaveBeenCalledTimes(1);
-			expect(svc.waitForServices).toHaveBeenCalledWith("math", 0);
-
-			return broker.stop();
-		});
-	});
-
-	describe("with multi dependency & $dependencyTimeout ", () => {
-
-		let broker = new ServiceBroker({ logger: false });
-
-		let schema = {
-			name: "simple",
-			dependencies: [
-				{ name: "math" },
-				{ name: "test", version: 2 }
-			],
-			settings: {
-				$dependencyTimeout: 5000
-			},
-			started: jest.fn()
-		};
-		let svc;
-
-		it("should called created", () => {
-			svc = broker.createService(schema);
-			svc.waitForServices = jest.fn(() => Promise.resolve());
-			return broker.start();
-		});
-
-		it("should called started", () => {
-			expect(schema.started).toHaveBeenCalledTimes(1);
-			expect(svc.waitForServices).toHaveBeenCalledTimes(1);
-			expect(svc.waitForServices).toHaveBeenCalledWith(schema.dependencies, 5000);
-
-			return broker.stop();
-		});
-	});
-
-});
-
-
-describe("Test applyMixins", () => {
-	let mixin1 = { name: "mixin1" };
-	let mixin2 = { name: "mixin2" };
-	let mixin3 = { name: "mixin3", mixins: mixin2 };
-
-	const oldMerge = Service.mergeSchemas;
-	beforeAll(() => Service.mergeSchemas = jest.fn(s1 => s1));
-	afterAll(() => Service.mergeSchemas = oldMerge);
-
-	it("should call Service.mergeSchemas with mixins", () => {
-		Service.mergeSchemas.mockClear();
-		let schema = {
-			name: "posts",
-			mixins: [
-				mixin1,
-				mixin2
-			]
-		};
-		Service.applyMixins(schema);
-
-		expect(Service.mergeSchemas).toHaveBeenCalledTimes(2);
-		expect(Service.mergeSchemas).toHaveBeenCalledWith(mixin2, mixin1);
-		expect(Service.mergeSchemas).toHaveBeenCalledWith(mixin2, schema);
-	});
-
-	it("should call Service.mergeSchemas with mixin", () => {
-		Service.mergeSchemas.mockClear();
-		let schema = {
-			name: "posts",
-			mixins: mixin2
-		};
-		Service.applyMixins(schema);
-
-		expect(Service.mergeSchemas).toHaveBeenCalledTimes(1);
-		expect(Service.mergeSchemas).toHaveBeenCalledWith(mixin2, schema);
-	});
-
-	it("should call Service.mergeSchemas with two level mixins", () => {
-		Service.mergeSchemas.mockClear();
-		let schema = {
-			name: "posts",
-			mixins: mixin3
-		};
-		Service.applyMixins(schema);
-
-		expect(Service.mergeSchemas).toHaveBeenCalledTimes(2);
-		expect(Service.mergeSchemas).toHaveBeenCalledWith(mixin2, mixin3);
-		expect(Service.mergeSchemas).toHaveBeenCalledWith(mixin2, schema);
-	});
-});
-
-describe("Test mergeSchemas", () => {
-
-	it("should merge two schemas", () => {
-
-		let origSchema = {
-			name: "posts",
-			settings: {
-				a: 5,
-				b: "10",
-				nested: {
-					id: 10
-				},
-				array: [
-					"first"
-				]
-			},
-
-			dependencies: [
-				"posts",
-				"users"
-			],
-
-			metadata: {
-				a: "a",
-				b: 33,
-				d: true,
-				nested: {
-					old: 1,
-					tag: "old"
+		it("should call the handler via this.events.xy", () => {
+			const schema = {
+				name: "posts",
+				events: {
+					"user.created": jest.fn(() => "User created"),
 				}
-			},
+			};
+			svc.parseServiceSchema(schema);
 
-			actions: {
-				get() {},
-				find() {},
-				list: {
-					cache: {
-						keys: ["id"]
+			const res = svc.events["user.created"]({ id: 5 });
+
+			expect(res).toBe("User created");
+			expect(schema.events["user.created"]).toBeCalledTimes(1);
+			expect(schema.events["user.created"]).toBeCalledWith(expect.any(Context));
+			expect(schema.events["user.created"].mock.calls[0][0].params).toEqual({ id: 5 });
+			expect(schema.events["user.created"].mock.calls[0][0].eventName).toBe("user.created");
+			expect(schema.events["user.created"].mock.calls[0][0].eventType).toBe("emit");
+			expect(schema.events["user.created"].mock.calls[0][0].eventGroups).toEqual(["posts"]);
+
+			schema.events["user.created"].mockClear();
+			const fakeCtx = { id: 123 };
+			const res2 = svc.events["user.created"]({ a: 5 }, { ctx: fakeCtx });
+
+			expect(res2).toBe("User created");
+			expect(schema.events["user.created"]).toBeCalledTimes(1);
+			expect(schema.events["user.created"]).toBeCalledWith(fakeCtx);
+		});
+
+	});
+
+	describe("Test _getPublicSettings", () => {
+
+		const broker = new ServiceBroker({ logger: false });
+
+		const svc = new Service(broker);
+
+		it("should skip the properties by $secureSettings", () => {
+			expect(svc._getPublicSettings({
+				a: 5,
+				auth: {
+					user: "user",
+					pass: "pass"
+				},
+				tokens: {
+					a: 111,
+					b: 222
+				},
+				x: "x",
+
+				$secureSettings: ["auth.pass", "tokens"]
+			})).toEqual({
+				a: 5,
+				auth: {
+					user: "user"
+				},
+				x: "x"
+			});
+		});
+
+		it("should keep all properties without $secureSettings", () => {
+			expect(svc._getPublicSettings({
+				a: 5,
+				auth: {
+					user: "user",
+					pass: "pass"
+				},
+				tokens: {
+					a: 111,
+					b: 222
+				},
+				x: "x"
+			})).toEqual({
+				a: 5,
+				auth: {
+					user: "user",
+					pass: "pass"
+				},
+				tokens: {
+					a: 111,
+					b: 222
+				},
+				x: "x"
+			});
+		});
+	});
+
+	describe("Test _init", () => {
+
+		const broker = new ServiceBroker({ logger: false });
+
+		const svc = new Service(broker);
+
+		jest.spyOn(broker, "addLocalService");
+		jest.spyOn(broker, "callMiddlewareHookSync");
+
+		it("should call broker methods", () => {
+			svc.parseServiceSchema({
+				name: "posts"
+			});
+
+			broker.addLocalService.mockClear();
+			broker.callMiddlewareHookSync.mockClear();
+
+			svc._init();
+
+			expect(broker.addLocalService).toBeCalledTimes(1);
+			expect(broker.addLocalService).toBeCalledWith(svc);
+
+			expect(broker.callMiddlewareHookSync).toBeCalledTimes(1);
+			expect(broker.callMiddlewareHookSync).toBeCalledWith("serviceCreated", [svc]);
+		});
+
+		it("should call service single 'created' hook", () => {
+			const created = jest.fn();
+			svc.parseServiceSchema({
+				name: "posts",
+				created
+			});
+
+			created.mockClear();
+
+			svc._init();
+
+			expect(created).toBeCalledTimes(1);
+			expect(created).toBeCalledWith();
+		});
+
+		it("should call service multi 'created' hook", () => {
+			let FLOW = [];
+			const created1 = jest.fn(() => FLOW.push("C1"));
+			const created2 = jest.fn(() => FLOW.push("C2"));
+			svc.parseServiceSchema({
+				name: "posts",
+				created: [created1, created2]
+			});
+
+			created1.mockClear();
+			created2.mockClear();
+			FLOW.length = 0;
+
+			svc._init();
+
+			expect(created1).toBeCalledTimes(1);
+			expect(created1).toBeCalledWith();
+			expect(created2).toBeCalledTimes(1);
+			expect(created2).toBeCalledWith();
+
+			expect(FLOW.join("-")).toBe("C1-C2");
+		});
+
+	});
+
+	describe("Test _start", () => {
+
+		const broker = new ServiceBroker({ logger: false });
+
+		const svc = new Service(broker);
+
+		jest.spyOn(broker, "registerLocalService");
+		jest.spyOn(broker, "callMiddlewareHook");
+		jest.spyOn(svc, "waitForServices").mockImplementation(async () => {});
+
+		it("should call broker methods", async () => {
+			svc.parseServiceSchema({
+				name: "posts"
+			});
+
+			broker.registerLocalService.mockClear();
+			broker.callMiddlewareHook.mockClear();
+			svc.waitForServices.mockClear();
+
+			await svc._start();
+
+			expect(broker.callMiddlewareHook).toBeCalledTimes(2);
+			expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(1, "serviceStarting", [svc]);
+			expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(2, "serviceStarted", [svc]);
+
+			expect(broker.registerLocalService).toBeCalledTimes(1);
+			expect(broker.registerLocalService).toBeCalledWith(svc._serviceSpecification);
+
+			expect(svc.waitForServices).toBeCalledTimes(0);
+		});
+
+		it("should call waitForServices if dependencies are defined", async () => {
+			svc.parseServiceSchema({
+				name: "posts",
+				dependencies: ["users", "auth"]
+			});
+
+			broker.registerLocalService.mockClear();
+			broker.callMiddlewareHook.mockClear();
+			svc.waitForServices.mockClear();
+
+			await svc._start();
+
+			expect(broker.callMiddlewareHook).toBeCalledTimes(2);
+			expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(1, "serviceStarting", [svc]);
+			expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(2, "serviceStarted", [svc]);
+
+			expect(broker.registerLocalService).toBeCalledTimes(1);
+			expect(broker.registerLocalService).toBeCalledWith(svc._serviceSpecification);
+
+			expect(svc.waitForServices).toBeCalledTimes(1);
+			expect(svc.waitForServices).toBeCalledWith(["users", "auth"], 0);
+		});
+
+		it("should call waitForServices if dependencies are defined & $dependencyTimeout", async () => {
+			svc.parseServiceSchema({
+				name: "posts",
+				settings: {
+					$dependencyTimeout: 3000
+				},
+				dependencies: ["users", "auth"]
+			});
+
+			svc.waitForServices.mockClear();
+
+			await svc._start();
+
+			expect(svc.waitForServices).toBeCalledTimes(1);
+			expect(svc.waitForServices).toBeCalledWith(["users", "auth"], 3000);
+		});
+
+		it("should call single started lifecycle event handler", async () => {
+			const started = jest.fn(async () => {});
+			svc.parseServiceSchema({
+				name: "posts",
+				started
+			});
+
+			await svc._start();
+
+			expect(started).toBeCalledTimes(1);
+			expect(started).toBeCalledWith();
+		});
+
+		it("should call multiple started lifecycle event handler", async () => {
+			let FLOW = [];
+
+			const started1 = jest.fn(async () => FLOW.push("S1"));
+			const started2 = jest.fn(async () => FLOW.push("S2"));
+			svc.parseServiceSchema({
+				name: "posts",
+				started: [started1, started2]
+			});
+
+			await svc._start();
+
+			expect(started1).toBeCalledTimes(1);
+			expect(started1).toBeCalledWith();
+
+			expect(started2).toBeCalledTimes(1);
+			expect(started2).toBeCalledWith();
+
+			expect(FLOW.join("-")).toBe("S1-S2");
+		});
+
+	});
+
+	describe("Test _stop", () => {
+		const broker = new ServiceBroker({ logger: false });
+
+		const svc = new Service(broker);
+
+		jest.spyOn(broker, "callMiddlewareHook");
+
+		it("should call broker methods", async () => {
+			svc.parseServiceSchema({
+				name: "posts"
+			});
+
+			broker.callMiddlewareHook.mockClear();
+
+			await svc._stop();
+
+			expect(broker.callMiddlewareHook).toBeCalledTimes(2);
+			expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(1, "serviceStopping", [svc], { reverse: true });
+			expect(broker.callMiddlewareHook).toHaveBeenNthCalledWith(2, "serviceStopped", [svc], { reverse: true });
+		});
+
+		it("should call single stopped lifecycle event handler", async () => {
+			const stopped = jest.fn(async () => {});
+			svc.parseServiceSchema({
+				name: "posts",
+				stopped
+			});
+
+			await svc._stop();
+
+			expect(stopped).toBeCalledTimes(1);
+			expect(stopped).toBeCalledWith();
+		});
+
+		it("should call multiple stopped lifecycle event handler", async () => {
+			let FLOW = [];
+			const stopped1 = jest.fn(async () => FLOW.push("S1"));
+			const stopped2 = jest.fn(async () => FLOW.push("S2"));
+			svc.parseServiceSchema({
+				name: "posts",
+				stopped: [stopped1, stopped2]
+			});
+
+			await svc._stop();
+
+			expect(stopped1).toBeCalledTimes(1);
+			expect(stopped1).toBeCalledWith();
+
+			expect(stopped2).toBeCalledTimes(1);
+			expect(stopped2).toBeCalledWith();
+
+			expect(FLOW.join("-")).toBe("S2-S1");
+		});
+	});
+
+	describe("Test _createAction", () => {
+		const broker = new ServiceBroker({ logger: false });
+
+		const svc = new Service(broker);
+		svc.parseServiceSchema({ name: "posts", version: 2 });
+
+		jest.spyOn(svc.Promise, "method");
+
+		it("should throw error if action schema is invalid", () => {
+			expect(() => { svc._createAction(null, "list"); }).toThrowError("Invalid action definition in 'list' action in 'v2.posts' service!");
+			expect(() => { svc._createAction("schema", "list"); }).toThrowError("Invalid action definition in 'list' action in 'v2.posts' service!");
+		});
+
+		it("should throw error if action handler is not defined", () => {
+			expect(() => { svc._createAction({}, "list"); })
+				.toThrowError("Missing action handler on 'list' action in 'v2.posts' service!");
+			expect(() => { svc._createAction({ handler: null }, "list"); })
+				.toThrowError("Missing action handler on 'list' action in 'v2.posts' service!");
+			expect(() => { svc._createAction({ handler: "wrong" }, "list"); })
+				.toThrowError("Missing action handler on 'list' action in 'v2.posts' service!");
+		});
+
+		it("should create action definition from a shorthand handler", () => {
+			const handler = jest.fn(function() {
+				expect(this).toBe(svc);
+				return "Hello";
+			});
+			svc.Promise.method.mockClear();
+
+			const res = svc._createAction(handler, "list");
+
+			expect(res).toEqual({
+				name: "v2.posts.list",
+				rawName: "list",
+				handler: expect.any(Function),
+				service: svc
+			});
+
+			expect.assertions(4);
+
+			expect(svc.Promise.method).toBeCalledTimes(1);
+
+			return res.handler().then(res => expect(res).toBe("Hello"));
+		});
+
+		it("should create action definition with cache", () => {
+			const handler = jest.fn(() => "Hello");
+			const res = svc._createAction({
+				params: {
+					a: "string"
+				},
+				cache: true,
+				etc: "etc",
+				handler
+			}, "list");
+
+			expect(res).toEqual({
+				name: "v2.posts.list",
+				rawName: "list",
+				handler: expect.any(Function),
+				service: svc,
+				cache: true,
+				etc: "etc",
+				params: {
+					a: "string"
+				}
+			});
+
+			expect.assertions(2);
+
+			return res.handler().then(res => expect(res).toBe("Hello"));
+		});
+
+		it("should create action definition without prefixes", () => {
+			svc.settings.$noServiceNamePrefix = true;
+
+			const handler = jest.fn(() => "Hello");
+			const res = svc._createAction({
+				handler
+			}, "list");
+
+			expect(res).toStrictEqual({
+				name: "list",
+				rawName: "list",
+				handler: expect.any(Function),
+				service: svc
+			});
+
+			svc.settings.$noServiceNamePrefix = false;
+		});
+
+		it("should set cache: true if $cache is defined in service settings", () => {
+			svc.settings.$cache = true;
+
+			const handler = jest.fn(() => "Hello");
+			const res = svc._createAction({
+				handler
+			}, "list");
+
+			expect(res).toStrictEqual({
+				name: "v2.posts.list",
+				rawName: "list",
+				handler: expect.any(Function),
+				service: svc,
+				cache: true
+			});
+		});
+
+		it("should set cache: true if $cache is defined in service settings", () => {
+			svc.settings.$cache = false;
+
+			const handler = jest.fn(() => "Hello");
+			const res = svc._createAction({
+				handler
+			}, "list");
+
+			expect(res).toStrictEqual({
+				name: "v2.posts.list",
+				rawName: "list",
+				handler: expect.any(Function),
+				service: svc,
+				cache: false
+			});
+		});
+
+		it("should not overwrite cache value if defined in action schema", () => {
+			svc.settings.$cache = true;
+
+			const handler = jest.fn(() => "Hello");
+			const res = svc._createAction({
+				handler,
+				cache: false
+			}, "list");
+
+			expect(res).toStrictEqual({
+				name: "v2.posts.list",
+				rawName: "list",
+				handler: expect.any(Function),
+				service: svc,
+				cache: false
+			});
+		});
+
+		it("should not overwrite cache value if defined in action schema", () => {
+			svc.settings.$cache = true;
+
+			const handler = jest.fn(() => "Hello");
+			const res = svc._createAction({
+				handler,
+				cache: {
+					keys: ["id"]
+				}
+			}, "list");
+
+			expect(res).toStrictEqual({
+				name: "v2.posts.list",
+				rawName: "list",
+				handler: expect.any(Function),
+				service: svc,
+				cache: {
+					keys: ["id"]
+				}
+			});
+		});
+	});
+
+	describe("Test _createEvent", () => {
+		const broker = new ServiceBroker({ logger: false });
+
+		const svc = new Service(broker);
+		svc.parseServiceSchema({ name: "posts", version: 2 });
+
+		jest.spyOn(svc.Promise, "method");
+
+		it("should throw error if event schema is invalid", () => {
+			expect(() => { svc._createEvent(null, "user.created"); }).toThrowError("Invalid event definition in 'user.created' event in 'v2.posts' service!");
+			expect(() => { svc._createEvent("schema", "user.created"); }).toThrowError("Invalid event definition in 'user.created' event in 'v2.posts' service!");
+		});
+
+		it("should throw error if event handler is not defined", () => {
+			expect(() => { svc._createEvent({}, "user.created"); })
+				.toThrowError("Missing event handler on 'user.created' event in 'v2.posts' service!");
+			expect(() => { svc._createEvent({ handler: null }, "user.created"); })
+				.toThrowError("Missing event handler on 'user.created' event in 'v2.posts' service!");
+			expect(() => { svc._createEvent({ handler: "wrong" }, "user.created"); })
+				.toThrowError("Missing event handler on 'user.created' event in 'v2.posts' service!");
+		});
+
+		it("should create event definition from a shorthand handler", () => {
+			const handler = jest.fn(function() {
+				expect(this).toBe(svc);
+				return "Hello";
+			});
+			const res = svc._createEvent(handler, "user.created");
+
+			expect(res).toEqual({
+				name: "user.created",
+				handler: expect.any(Function),
+				service: svc
+			});
+
+			expect.assertions(3);
+
+			return res.handler({}).then(res => expect(res).toBe("Hello"));
+		});
+
+		it("should create event definition from event schema", () => {
+			svc.Promise.method.mockClear();
+			const handler = jest.fn(function() {
+				expect(this).toBe(svc);
+				return "Hello";
+			});
+			const res = svc._createEvent({
+				name: "user.updated",
+				handler,
+				etc: "etc"
+			}, "other");
+
+			expect(res).toEqual({
+				name: "user.updated",
+				handler: expect.any(Function),
+				service: svc,
+				etc: "etc"
+			});
+
+			expect.assertions(4);
+
+			expect(svc.Promise.method).toBeCalledTimes(1);
+
+			return res.handler({}).then(res => expect(res).toBe("Hello"));
+		});
+
+		it("should create event definition with multiple handlers", () => {
+			const handler1 = jest.fn(function() {
+				expect(this).toBe(svc);
+				return "Hello1";
+			});
+			const handler2 = jest.fn(function() {
+				expect(this).toBe(svc);
+				return "Hello2";
+			});
+
+			const res = svc._createEvent({
+				handler: [handler1, handler2]
+			}, "user.updated");
+
+			expect(res).toEqual({
+				name: "user.updated",
+				handler: expect.any(Function),
+				service: svc
+			});
+
+			expect.assertions(4);
+
+			return res.handler({}).then(res => expect(res).toEqual(["Hello1", "Hello2"]));
+		});
+
+		it("should call handler with legacy arguments", () => {
+			const handler = function(payload, nodeID, eventName, ctx) {
+				expect(this).toBe(svc);
+				return { payload, nodeID, eventName, ctx };
+			};
+
+			const res = svc._createEvent({
+				handler
+			}, "user.updated");
+
+			expect.assertions(5);
+
+			const ctx = {
+				params: { a: 5 },
+				nodeID: "node-100",
+				eventName: "user.removed"
+			};
+			return res.handler(ctx).then(res => {
+				expect(res.payload).toEqual({ a: 5 });
+				expect(res.nodeID).toEqual("node-100");
+				expect(res.eventName).toEqual("user.removed");
+				expect(res.ctx).toBe(ctx);
+			});
+		});
+
+		it("should call handler with context", () => {
+			const handler = function(ctx) {
+				expect(this).toBe(svc);
+				return { ctx };
+			};
+
+			const res = svc._createEvent({
+				handler
+			}, "user.updated");
+
+			expect.assertions(5);
+
+			const ctx = {
+				params: { a: 5 },
+				nodeID: "node-100",
+				eventName: "user.removed"
+			};
+			return res.handler(ctx).then(res => {
+				expect(res.payload).toBeUndefined();
+				expect(res.nodeID).toBeUndefined();
+				expect(res.eventName).toBeUndefined();
+				expect(res.ctx).toBe(ctx);
+			});
+		});
+
+		it("should call handler with multiple times", () => {
+			const handler1 = function(ctx) {
+				expect(this).toBe(svc);
+				return { ctx };
+			};
+
+			const handler2 = function(payload, nodeID, eventName, ctx) {
+				expect(this).toBe(svc);
+				return { payload, nodeID, eventName, ctx };
+			};
+
+			const res = svc._createEvent({
+				handler: [handler1, handler2]
+			}, "user.updated");
+
+			//expect.assertions(5);
+
+			const ctx = {
+				params: { a: 5 },
+				nodeID: "node-100",
+				eventName: "user.removed"
+			};
+			return res.handler(ctx).then(([res1, res2]) => {
+				expect(res1.payload).toBeUndefined();
+				expect(res1.nodeID).toBeUndefined();
+				expect(res1.eventName).toBeUndefined();
+				expect(res1.ctx).toBe(ctx);
+
+				expect(res2.payload).toEqual({ a: 5 });
+				expect(res2.nodeID).toEqual("node-100");
+				expect(res2.eventName).toEqual("user.removed");
+				expect(res2.ctx).toBe(ctx);
+			});
+		});
+	});
+
+	describe("Test emitLocalEventHandler", () => {
+		const broker = new ServiceBroker({ logger: false });
+
+		const svc = new Service(broker);
+
+		it("should call the event handler", async () => {
+			svc.parseServiceSchema({
+				name: "posts",
+				version: 2,
+				events: {
+					"user.created"(ctx) {
+						expect(this).toBe(svc);
+						expect(ctx.params).toEqual({ id: 5 });
+						expect(ctx.eventName).toBe("user.created");
+						expect(ctx.eventGroups).toEqual(["posts"]);
+						expect(ctx.eventType).toBe("emit");
+						return "Hello";
+					}
+				}
+			});
+
+			const res = await svc.emitLocalEventHandler("user.created", { id: 5 });
+			expect(res).toBe("Hello");
+
+			expect.assertions(6);
+		});
+
+		it("should call the event handler with opts", async () => {
+			svc.parseServiceSchema({
+				name: "posts",
+				version: 2,
+				events: {
+					"user.created"(ctx) {
+						expect(this).toBe(svc);
+						expect(ctx.params).toEqual({ id: 5 });
+						expect(ctx.eventName).toBe("user.created");
+						expect(ctx.eventGroups).toEqual(["posts"]);
+						expect(ctx.eventType).toBe("emit");
+						expect(ctx.requestID).toBe("12345");
+						return "Hello";
+					}
+				}
+			});
+
+			const res = await svc.emitLocalEventHandler("user.created", { id: 5 }, { requestID: "12345" });
+			expect(res).toBe("Hello");
+
+			expect.assertions(7);
+		});
+
+		it("should throw error if event is not exist", async () => {
+			try {
+				await svc.emitLocalEventHandler("not.found");
+			} catch(err) {
+				expect(err.code).toBe(500);
+				expect(err.type).toBe("NOT_FOUND_EVENT");
+				expect(err.message).toBe("No 'not.found' registered local event handler");
+				expect(err.data).toEqual({ eventName: "not.found" });
+			}
+		});
+
+	});
+
+	describe("Test waitForServices", () => {
+		const broker = new ServiceBroker({ logger: false });
+		jest.spyOn(broker, "waitForServices").mockImplementation(async () => {});
+
+		const svc = new Service(broker);
+
+		it("should call broker.waitForServices", async () => {
+			await svc.waitForServices(["users", "auth"], 3000, 500);
+
+			expect(broker.waitForServices).toHaveBeenCalledTimes(1);
+			expect(broker.waitForServices).toHaveBeenCalledWith(["users", "auth"], 3000, 500, svc.logger);
+		});
+
+	});
+
+	describe("Test static applyMixins", () => {
+
+		beforeAll(() => jest.spyOn(Service, "mergeSchemas").mockImplementation(s => s));
+		afterAll(() => Service.mergeSchemas.mockRestore());
+
+		it("should return the schema if no mixins defined", () => {
+			Service.mergeSchemas.mockClear();
+			const schema = { name: "posts" };
+
+			const res = Service.applyMixins(schema);
+
+			expect(res).toBe(schema);
+			expect(Service.mergeSchemas).toBeCalledTimes(0);
+		});
+
+		it("should call mergeSchema once", () => {
+			Service.mergeSchemas.mockClear();
+			const mixin1 = {
+				name: "users"
+			};
+
+			const schema = {
+				name: "posts",
+				mixins: mixin1
+			};
+
+			const res = Service.applyMixins(schema);
+
+			expect(res).toEqual({
+				name: "users"
+			});
+			expect(Service.mergeSchemas).toBeCalledTimes(1);
+			expect(Service.mergeSchemas).toBeCalledWith(mixin1, schema);
+		});
+
+		it("should call mergeSchema twice", () => {
+			Service.mergeSchemas.mockClear();
+			const mixin1 = {
+				name: "users"
+			};
+
+			const mixin2 = {
+				version: 2
+			};
+
+			const schema = {
+				name: "posts",
+				mixins: [mixin1, mixin2]
+			};
+
+			const res = Service.applyMixins(schema);
+
+			expect(res).toEqual({
+				version: 2
+			});
+
+			expect(Service.mergeSchemas).toBeCalledTimes(2);
+			expect(Service.mergeSchemas).toHaveBeenNthCalledWith(1, mixin2, mixin1);
+			expect(Service.mergeSchemas).toHaveBeenNthCalledWith(2, mixin2, schema);
+		});
+
+		it("should call mergeSchema for multi-level mixins", () => {
+			Service.mergeSchemas.mockClear();
+			const mixin2 = {
+				version: 2
+			};
+
+			const mixin1 = {
+				name: "users",
+				mixins: [mixin2]
+			};
+
+			const schema = {
+				name: "posts",
+				mixins: [mixin1]
+			};
+
+			const res = Service.applyMixins(schema);
+
+			expect(res).toEqual({
+				version: 2
+			});
+
+			expect(Service.mergeSchemas).toBeCalledTimes(2);
+			expect(Service.mergeSchemas).toHaveBeenNthCalledWith(1, mixin2, mixin1);
+			expect(Service.mergeSchemas).toHaveBeenNthCalledWith(2, mixin2, schema);
+		});
+	});
+
+	describe("Test static mergeSchemas", () => {
+
+		beforeAll(() => {
+			jest.spyOn(Service, "mergeSchemaSettings").mockImplementation(s => s);
+			jest.spyOn(Service, "mergeSchemaMetadata").mockImplementation(s => s);
+			jest.spyOn(Service, "mergeSchemaHooks").mockImplementation(s => s);
+			jest.spyOn(Service, "mergeSchemaActions").mockImplementation(s => s);
+			jest.spyOn(Service, "mergeSchemaMethods").mockImplementation(s => s);
+			jest.spyOn(Service, "mergeSchemaEvents").mockImplementation(s => s);
+			jest.spyOn(Service, "mergeSchemaLifecycleHandlers").mockImplementation(s => s);
+			jest.spyOn(Service, "mergeSchemaUniqArray").mockImplementation(s => s);
+			jest.spyOn(Service, "mergeSchemaUnknown").mockImplementation(s => s);
+		});
+		afterAll(() => {
+			Service.mergeSchemaSettings.mockRestore();
+			Service.mergeSchemaMetadata.mockRestore();
+			Service.mergeSchemaHooks.mockRestore();
+			Service.mergeSchemaActions.mockRestore();
+			Service.mergeSchemaMethods.mockRestore();
+			Service.mergeSchemaEvents.mockRestore();
+			Service.mergeSchemaLifecycleHandlers.mockRestore();
+			Service.mergeSchemaUniqArray.mockRestore();
+			Service.mergeSchemaUnknown.mockRestore();
+		});
+
+		it("should call merge methods", () => {
+			const mixin2 = {};
+
+			const mixin = {
+				name: "posts",
+				version: 3,
+
+				mixins: [ mixin2 ],
+
+				dependencies: ["users", "auth"],
+
+				settings: {
+					a: 5
+				},
+
+				metadata: {
+					region: "us-west"
+				},
+
+				hooks: {
+					before: {
+						"*": () => {}
+					},
+					after: {
+						list: [
+							() => {},
+							() => {}
+						]
+					},
+					error: {
+						"*": () => {}
+					}
+				},
+
+				actions: {
+					find() {},
+
+					get: {
+						cache: true,
+						params: {
+							a: "string"
+						}
+					},
+
+					delete: false
+				},
+
+				methods: {
+					doSomething() {},
+				},
+
+				events: {
+					"user.created"() {},
+					"user.updated": {
+						group: "user",
+						handler: () => {}
+					}
+				},
+
+				created() {},
+				started() {},
+				stopped() {},
+
+				custom: "123"
+			};
+
+			Service.mergeSchemas({}, mixin);
+
+			expect(Service.mergeSchemaSettings).toBeCalledTimes(1);
+			expect(Service.mergeSchemaSettings).toBeCalledWith(mixin.settings, undefined);
+
+			expect(Service.mergeSchemaMetadata).toBeCalledTimes(1);
+			expect(Service.mergeSchemaMetadata).toBeCalledWith(mixin.metadata, undefined);
+
+			expect(Service.mergeSchemaHooks).toBeCalledTimes(1);
+			expect(Service.mergeSchemaHooks).toBeCalledWith(mixin.hooks, {});
+
+			expect(Service.mergeSchemaActions).toBeCalledTimes(1);
+			expect(Service.mergeSchemaActions).toBeCalledWith(mixin.actions, {});
+
+			expect(Service.mergeSchemaMethods).toBeCalledTimes(1);
+			expect(Service.mergeSchemaMethods).toBeCalledWith(mixin.methods, undefined);
+
+			expect(Service.mergeSchemaEvents).toBeCalledTimes(1);
+			expect(Service.mergeSchemaEvents).toBeCalledWith(mixin.events, {});
+
+			expect(Service.mergeSchemaLifecycleHandlers).toBeCalledTimes(3);
+			expect(Service.mergeSchemaLifecycleHandlers).toBeCalledWith(mixin.created, undefined);
+			expect(Service.mergeSchemaLifecycleHandlers).toBeCalledWith(mixin.started, undefined);
+			expect(Service.mergeSchemaLifecycleHandlers).toBeCalledWith(mixin.stopped, undefined);
+
+			expect(Service.mergeSchemaUniqArray).toBeCalledTimes(2);
+			expect(Service.mergeSchemaUniqArray).toHaveBeenNthCalledWith(1, mixin.mixins, undefined);
+			expect(Service.mergeSchemaUniqArray).toHaveBeenNthCalledWith(2, mixin.dependencies, undefined);
+
+			expect(Service.mergeSchemaUnknown).toBeCalledTimes(1);
+			expect(Service.mergeSchemaUnknown).toHaveBeenNthCalledWith(1, mixin.custom, undefined);
+
+		});
+
+		it("should call custom merge method", () => {
+			Service.mergeSchemaUnknown.mockClear();
+			Service.mergeSchemaMyProp = jest.fn();
+
+			const mixin = {
+				myProp: "123"
+			};
+
+			Service.mergeSchemas({}, mixin);
+
+			expect(Service.mergeSchemaMyProp).toBeCalledTimes(1);
+			expect(Service.mergeSchemaMyProp).toBeCalledWith("123", undefined);
+			expect(Service.mergeSchemaUnknown).toBeCalledTimes(0);
+		});
+
+		it("should not overwrite the name & version", () => {
+			Service.mergeSchemaUnknown.mockClear();
+
+			const mixed = Service.mergeSchemas({
+				name: "first",
+				version: 1
+			}, {});
+
+			expect(mixed).toEqual({
+				name: "first",
+				version: 1
+			});
+		});
+
+		it("should overwrite the name & version", () => {
+			Service.mergeSchemaUnknown.mockClear();
+
+			const mixed = Service.mergeSchemas({
+				name: "first",
+				version: 1
+			}, {
+				name: "second",
+				version: 2
+			});
+
+			expect(mixed).toEqual({
+				name: "second",
+				version: 2
+			});
+		});
+
+	});
+
+	describe("Test static mergeSchemaSettings", () => {
+
+		it("should merge values", () => {
+			const src = {
+				a: 5,
+				b: "John",
+				c: {
+					d: true,
+					e: 45.8
+				}
+			};
+
+			const prev = {
+				a: 10,
+				c: {
+					e: 10.5,
+					f: "Fox"
+				}
+			};
+
+			const res = Service.mergeSchemaSettings(src, prev);
+			expect(res).toEqual({
+				a: 5,
+				b: "John",
+				c: {
+					d: true,
+					e: 45.8,
+					f: "Fox"
+				}
+			});
+		});
+
+		it("should merge values with $secureSettings", () => {
+			const src = {
+				$secureSettings: ["a", "b"],
+				a: 5,
+				b: "John",
+				c: {
+					d: true,
+					e: 45.8
+				}
+			};
+
+			const prev = {
+				$secureSettings: ["c", "b"],
+				a: 10,
+				c: {
+					e: 10.5,
+					f: "Fox"
+				}
+			};
+
+			const res = Service.mergeSchemaSettings(src, prev);
+			expect(res).toEqual({
+				$secureSettings: ["a", "b", "c"],
+				a: 5,
+				b: "John",
+				c: {
+					d: true,
+					e: 45.8,
+					f: "Fox"
+				}
+			});
+		});
+	});
+
+	describe("Test static mergeSchemaMetadata", () => {
+
+		it("should merge values", () => {
+			const src = {
+				a: 5,
+				b: "John",
+				c: {
+					d: true,
+					e: 45.8
+				}
+			};
+
+			const prev = {
+				a: 10,
+				c: {
+					e: 10.5,
+					f: "Fox"
+				}
+			};
+
+			const res = Service.mergeSchemaMetadata(src, prev);
+			expect(res).toEqual({
+				a: 5,
+				b: "John",
+				c: {
+					d: true,
+					e: 45.8,
+					f: "Fox"
+				}
+			});
+		});
+	});
+
+	describe("Test static mergeSchemaUniqArray", () => {
+
+		it("should merge values", () => {
+			const src = [1,2,3,4,5];
+			const prev = [2,4,6,8,10];
+
+			const res = Service.mergeSchemaUniqArray(src, prev);
+			expect(res).toEqual([1, 2, 3, 4, 5, 6, 8, 10]);
+		});
+
+		it("should merge objects", () => {
+			const src = [
+				{ id: 1 },
+				{ id: 2 },
+				{ id: 3 },
+				{ id: 4 },
+				{ id: 5 },
+			];
+			const prev = [
+				{ id: 2 },
+				{ id: 4 },
+				{ id: 6 },
+				{ id: 8 },
+				{ id: 10 },
+			];
+
+			const res = Service.mergeSchemaUniqArray(src, prev);
+			expect(res).toEqual([
+				{ id: 1 },
+				{ id: 2 },
+				{ id: 3 },
+				{ id: 4 },
+				{ id: 5 },
+				{ id: 6 },
+				{ id: 8 },
+				{ id: 10 }
+			]);
+		});
+	});
+
+	describe("Test static mergeSchemaDependencies", () => {
+
+		it("should merge values", () => {
+			jest.spyOn(Service, "mergeSchemaUniqArray");
+
+			const src = [1,2,3,4,5];
+			const prev = [2,4,6,8,10];
+
+			const res = Service.mergeSchemaDependencies(src, prev);
+			expect(res).toEqual([1, 2, 3, 4, 5, 6, 8, 10]);
+
+			expect(Service.mergeSchemaUniqArray).toBeCalledTimes(1);
+			expect(Service.mergeSchemaUniqArray).toBeCalledWith(src, prev);
+		});
+	});
+
+	describe("Test static mergeSchemaHooks", () => {
+
+		it("should merge values", () => {
+			const src = {
+				before: {
+					all: "src-before-all"
+				},
+				after: {
+					list: [
+						"src-after-list1",
+						"src-after-list2",
+					]
+				},
+				error: {
+					all: "src-error-all",
+					remove: "src-error-remove",
+				}
+			};
+
+			const prev = {
+				before: {
+					all: "prev-before-all",
+					create: "prev-before-create"
+				},
+				after: {
+					list: "prev-after-list"
+				},
+				error: {
+					all: "prev-error-all"
+				}
+			};
+
+			const res = Service.mergeSchemaHooks(src, prev);
+			expect(res).toEqual({
+				before: {
+					all: ["prev-before-all", "src-before-all"],
+					create: "prev-before-create"
+				},
+				after: {
+					list: ["src-after-list1", "src-after-list2", "prev-after-list"]
+				},
+				error: {
+					all: ["src-error-all", "prev-error-all"],
+					remove: ["src-error-remove"]
+				}
+			});
+		});
+	});
+
+	describe("Test static mergeSchemaActions", () => {
+
+		it("should merge actions", () => {
+			const src = {
+				create() {},
+				find: {
+					params: {
+						count: "number"
 					},
 					handler() {}
 				},
-				create() {},
-				update: {
-					use: [],
-					handler() {
+				update: false
+			};
 
-					}
-				},
-				clean() {}
-			},
-
-			events: {
-				"created"() {},
-				"updated"() {},
-				"removed": {
-					group: "mail",
-					handler: () => {}
-				},
-				"inserted"() {}
-			},
-
-			methods: {
-				getByID() {},
-				notify() {}
-			},
-
-			created: jest.fn(),
-			started: jest.fn(),
-			stopped: jest.fn()
-		};
-
-		let newSchema = {
-			name: "users",
-			version: 2,
-			settings: {
-				b: 100,
-				c: true,
-				nested: {
-					name: "John"
-				},
-				array: [
-					"second",
-					"third"
-				]
-			},
-
-			dependencies: "math",
-
-			metadata: {
-				a: "aaa",
-				b: 25,
-				c: "metadata",
-				nested: {
-					tag: "new",
-					res: "test"
-				}
-			},
-
-			actions: {
+			const prev = {
 				find: {
-					cache: false,
+					cache: true
+				},
+
+				update: {
 					handler() {}
 				},
-				list() {},
+
+				patch: {
+					params: {
+						update: "object"
+					},
+					handler() {}
+				},
+
+				remove() {}
+			};
+
+			const res = Service.mergeSchemaActions(src, prev);
+			expect(res).toEqual({
 				create: {
-					cache: {
-						keys: ["id"]
+					handler: expect.any(Function)
+				},
+				find: {
+					cache: true,
+					params: {
+						count: "number"
+					},
+					handler: expect.any(Function)
+				},
+				patch: {
+					params: {
+						update: "object"
+					},
+					handler: expect.any(Function)
+				},
+				remove: expect.any(Function),
+			});
+		});
+
+		it("should merge actions with hooks", () => {
+			const src = {
+				create: {
+					hooks: {
+						before: "src-create-before",
+						after: "src-create-after",
+						error: "src-create-error"
 					}
 				},
-				update() {
+				find: {
+					hooks: {
+						before: "src-create-before",
+						after: ["src-create-after1", "src-create-after2"],
+						error: "src-create-error",
+					},
+					handler() {}
 				},
-				remove() {},
-				clean: false
-			},
+				update: {
+					handler() {}
+				}
+			};
 
-			events: {
-				"created"() {},
-				"removed"() {},
-				"cleared"() {},
-				"inserted": {
-					group: "payment",
-					handler: () => {}
+			const prev = {
+				create: {
+					handler() {}
 				},
-			},
 
-			methods: {
-				getByID() {},
-				checkPermission() {}
-			},
+				find: {
+					hooks: {
+						before: "prev-create-before",
+						after: ["prev-create-after1", "prev-create-after2"],
+						error: "prev-create-error",
+					},
+				},
 
-			created: jest.fn(),
-			started: jest.fn(),
-			stopped: jest.fn(),
+				update: {
+					hooks: {
+						before: "prev-update-before",
+						after: "prev-update-after",
+						error: "prev-update-error"
+					}
+				},
+			};
 
-			customProp: "test"
-		};
-
-		let res = Service.mergeSchemas(origSchema, newSchema);
-
-		expect(res).toBeDefined();
-		expect(res.name).toBe("users");
-		expect(res.version).toBe(2);
-		expect(res.settings).toEqual({
-			a: 5,
-			b: 100,
-			c: true,
-			nested: {
-				id: 10,
-				name: "John"
-			},
-			array: [
-				"second",
-				"third"
-			]
+			const res = Service.mergeSchemaActions(src, prev);
+			expect(res).toEqual({
+				create: {
+					hooks: {
+						before: "src-create-before",
+						after: "src-create-after",
+						error: "src-create-error"
+					},
+					handler: expect.any(Function)
+				},
+				find: {
+					hooks: {
+						before: ["prev-create-before", "src-create-before"],
+						after: ["src-create-after1", "src-create-after2", "prev-create-after1", "prev-create-after2"],
+						error: ["src-create-error", "prev-create-error"]
+					},
+					handler: expect.any(Function)
+				},
+				update: {
+					hooks: {
+						before: "prev-update-before",
+						after: "prev-update-after",
+						error: "prev-update-error"
+					},
+					handler: expect.any(Function)
+				},
+			});
 		});
-		expect(res.metadata).toEqual({
-			a: "aaa",
-			b: 25,
-			c: "metadata",
-			d: true,
-			nested: {
-				old: 1,
-				tag: "new",
-				res: "test"
-			}
-		});
-		expect(res.dependencies).toEqual(["math", "posts", "users"]);
-
-		// Actions
-		expect(res.actions.get).toBe(origSchema.actions.get);
-		expect(res.actions.find.handler).toBe(newSchema.actions.find.handler);
-		expect(res.actions.find.cache).toBe(false);
-		expect(res.actions.list.handler).toBe(newSchema.actions.list);
-		expect(res.actions.remove.handler).toBe(newSchema.actions.remove);
-		expect(res.actions.clean).toBeUndefined();
-
-		// Merge action definition
-		expect(res.actions.create).toEqual({
-			cache: {
-				keys: ["id"]
-			},
-			handler: origSchema.actions.create
-		});
-		expect(res.actions.create.handler).toBe(origSchema.actions.create);
-
-		expect(res.actions.update).toEqual({
-			use: [],
-			handler: newSchema.actions.update
-		});
-		expect(res.actions.update.handler).toBe(newSchema.actions.update);
-
-		// Events
-		expect(res.events.created.handler).toBeInstanceOf(Array);
-		expect(res.events.created.handler[0]).toBe(origSchema.events.created);
-		expect(res.events.created.handler[1]).toBe(newSchema.events.created);
-
-		expect(res.events.updated).toBe(origSchema.events.updated);
-
-		expect(res.events.removed.handler).toBeInstanceOf(Array);
-		expect(res.events.removed.handler[0]).toBe(origSchema.events.removed.handler);
-		expect(res.events.removed.handler[1]).toBe(newSchema.events.removed);
-		expect(res.events.removed.group).toBe("mail");
-
-		expect(res.events.cleared.handler).toBe(newSchema.events.cleared);
-
-		expect(res.events.inserted.handler).toBeInstanceOf(Array);
-		expect(res.events.inserted.handler[0]).toBe(origSchema.events.inserted);
-		expect(res.events.inserted.handler[1]).toBe(newSchema.events.inserted.handler);
-		expect(res.events.inserted.group).toBe("payment");
-
-		// Methods
-		expect(res.methods.getByID).toBe(newSchema.methods.getByID);
-		expect(res.methods.notify).toBe(origSchema.methods.notify);
-		expect(res.methods.checkPermission).toBe(newSchema.methods.checkPermission);
-
-		// Lifecycle handlers
-		expect(res.created).toBeInstanceOf(Array);
-		expect(res.started).toBeInstanceOf(Array);
-		expect(res.stopped).toBeInstanceOf(Array);
-
-		expect(res.created[0]).toBe(origSchema.created);
-		expect(res.created[1]).toBe(newSchema.created);
-
-		expect(res.started[0]).toBe(origSchema.started);
-		expect(res.started[1]).toBe(newSchema.started);
-
-		expect(res.stopped[0]).toBe(origSchema.stopped);
-		expect(res.stopped[1]).toBe(newSchema.stopped);
-
-		expect(res.customProp).toBe("test");
-
 	});
 
-	it("should concat lifecycle events", () => {
+	describe("Test static mergeSchemaMethods", () => {
 
-		let origSchema = {
-			created: jest.fn(),
-			started: jest.fn()
-		};
+		it("should merge values", () => {
+			const src = {
+				find: "src-find",
+				list: "src-list",
+				update: "src-update",
+			};
 
-		let newSchema = {
-			created: jest.fn(),
-			stopped: jest.fn()
-		};
+			const prev = {
+				create: "prev-create",
+				list: "prev-list",
+				update: "prev-update",
+				remove: "prev-remove",
+			};
 
-		let res = Service.mergeSchemas(origSchema, newSchema);
-
-		expect(res).toBeDefined();
-
-		expect(res.created).toBeInstanceOf(Array);
-		expect(res.created.length).toBe(2);
-		expect(res.created[0]).toBe(origSchema.created);
-		expect(res.created[1]).toBe(newSchema.created);
-
-		expect(res.started).toBe(origSchema.started);
-
-		expect(res.stopped).toBeInstanceOf(Array);
-		expect(res.stopped.length).toBe(1);
-		expect(res.stopped[0]).toBe(newSchema.stopped);
-
+			const res = Service.mergeSchemaMethods(src, prev);
+			expect(res).toEqual({
+				create: "prev-create",
+				find: "src-find",
+				list: "src-list",
+				update: "src-update",
+				remove: "prev-remove",
+			});
+		});
 	});
 
-	describe("merge schemas with unique dependencies", () => {
-		it("should merge dependencies defined in shorthand notation", () => {
-			const origSchema = {
-				dependencies: ["users", "mail"]
-			};
-			const mixinSchema = {
-				dependencies: ["users", "posts"]
+	describe("Test static mergeSchemaEvents", () => {
+
+		it("should merge actions", () => {
+			const src = {
+				create() {},
+				find: {
+					params: {
+						count: "number"
+					},
+					handler() {}
+				},
+				update: {
+					handler() {}
+				},
 			};
 
-			const result = Service.mergeSchemas(origSchema, mixinSchema);
-			expect(result.dependencies).toEqual(["users", "posts", "mail"]);
+			const prev = {
+				find: {
+					group: "user",
+				},
+
+				update: {
+					handler() {}
+				},
+
+				patch: {
+					params: {
+						update: "object"
+					},
+					handler() {}
+				},
+
+				remove() {}
+			};
+
+			const res = Service.mergeSchemaEvents(src, prev);
+			expect(res).toEqual({
+				create: {
+					handler: expect.any(Function)
+				},
+				find: {
+					group: "user",
+					params: {
+						count: "number"
+					},
+					handler: expect.any(Function)
+				},
+				update: {
+					handler: [expect.any(Function), expect.any(Function)]
+				},
+				patch: {
+					params: {
+						update: "object"
+					},
+					handler: expect.any(Function)
+				},
+				remove: expect.any(Function),
+			});
 		});
+	});
 
-		it("should merge versioned dependencies", () => {
-			const origSchema = {
-				dependencies: [{ name: "users", version: 1 }, { name: "mail", version: "staging" }]
-			};
-			const mixinSchema = {
-				dependencies: [{ name: "users", version: 1 }, { name: "posts", version: "v2" }]
+	describe("Test static mergeSchemaLifecycleHandlers", () => {
+
+		it("should merge values", () => {
+			const src = {
+				created: "src-created",
+				started: "src-started",
+				stopped: "src-stopped",
 			};
 
-			const result = Service.mergeSchemas(origSchema, mixinSchema);
-			expect(result.dependencies).toEqual([{ name: "users", version: 1 }, { name: "posts", version: "v2" }, { name: "mail", version: "staging" }]);
+			const prev = {
+				created: "prev-created",
+				started: "prev-started",
+				stopped: "prev-stopped",
+			};
+
+			const res = Service.mergeSchemaLifecycleHandlers("src-created", "prev-created");
+			expect(res).toEqual(["prev-created", "src-created"]);
 		});
+	});
 
-		it("should merge mixed", () => {
-			const origSchema = {
-				dependencies: [{ name: "users", version: 1 }, "mail"]
-			};
-			const mixinSchema = {
-				dependencies: ["users", { name: "posts", version: "v2" }]
-			};
+	describe("Test static mergeSchemaUnknown", () => {
 
-			const result = Service.mergeSchemas(origSchema, mixinSchema);
-			expect(result.dependencies).toEqual(["users", { name: "posts", version: "v2" }, { name: "users", version: 1 }, "mail"]);
+		it("should merge values", () => {
+			expect(Service.mergeSchemaUnknown("John", "Bob")).toBe("John");
+			expect(Service.mergeSchemaUnknown("John", null)).toBe("John");
+			expect(Service.mergeSchemaUnknown(null, "Bob")).toBeNull();
+			expect(Service.mergeSchemaUnknown(null, null)).toBeNull();
+
+			expect(Service.mergeSchemaUnknown("John", undefined)).toBe("John");
+			expect(Service.mergeSchemaUnknown(undefined, "Bob")).toBe("Bob");
+			expect(Service.mergeSchemaUnknown(undefined, undefined)).toBeUndefined();
 		});
+	});
 
+	describe("Test static getVersionedFullName", () => {
+
+		it("should create version service names", () => {
+			expect(Service.getVersionedFullName("posts")).toBe("posts");
+			expect(Service.getVersionedFullName("posts", 5)).toBe("v5.posts");
+			expect(Service.getVersionedFullName("posts", "testing")).toBe("testing.posts");
+		});
 	});
 
 });
 
-
-describe("Test $secureSettings", () => {
-
-	let schema = {
-		name: "demo",
-		settings: {
-			jwtSecret: "12345",
-			other: "some",
-			topSecret: {
-				name: "John",
-				password: "password",
-				age: 33
-			}
-		},
-	};
-
-	let broker;
-
-	function createService(schema) {
-		broker = new ServiceBroker({ logger: false });
-		const svc = broker.createService(schema);
-		return broker.start().then(() => svc);
-	}
-
-	afterEach(() => broker ? broker.stop() : null);
-
-	it("should leave all settings", () => {
-		return createService(schema).catch(protectReject).then(svc => {
-			expect(svc.settings).toEqual({
-				jwtSecret: "12345",
-				other: "some",
-				topSecret: {
-					name: "John",
-					password: "password",
-					age: 33
-				}
-			});
-
-			const list = broker.registry.services.list({ onlyLocal: true, skipInternal: true });
-			expect(list[0].settings).toEqual({
-				jwtSecret: "12345",
-				other: "some",
-				topSecret: {
-					name: "John",
-					password: "password",
-					age: 33
-				}
-			});
-		});
-	});
-
-	it("should omit protected settings", () => {
-		schema.settings.$secureSettings = ["jwtSecret", "topSecret"];
-
-		return createService(schema).catch(protectReject).then(svc => {
-			expect(svc.settings).toEqual({
-				$secureSettings: ["jwtSecret", "topSecret"],
-				jwtSecret: "12345",
-				other: "some",
-				topSecret: {
-					name: "John",
-					password: "password",
-					age: 33
-				}
-			});
-
-			const list = broker.registry.services.list({ onlyLocal: true, skipInternal: true });
-			expect(list[0].settings).toEqual({
-				other: "some"
-			});
-		});
-	});
-
-	it("should omit protected nested settings", () => {
-		schema.settings.$secureSettings = ["jwtSecret", "topSecret.password"];
-
-		return createService(schema).catch(protectReject).then(svc => {
-			expect(svc.settings).toEqual({
-				$secureSettings: ["jwtSecret", "topSecret.password"],
-				jwtSecret: "12345",
-				other: "some",
-				topSecret: {
-					name: "John",
-					password: "password",
-					age: 33
-				}
-			});
-
-			const list = broker.registry.services.list({ onlyLocal: true, skipInternal: true });
-			expect(list[0].settings).toEqual({
-				other: "some",
-				topSecret: {
-					name: "John",
-					age: 33
-				}
-			});
-		});
-	});
-
-	it("should merge $secureSettings from mixins & omit protected settings", () => {
-		schema.mixins = [{
-			settings: {
-				$secureSettings: ["topSecret.password"]
-			}
-		}];
-		schema.settings.$secureSettings = ["jwtSecret"];
-
-		return createService(schema).catch(protectReject).then(svc => {
-			expect(svc.settings).toEqual({
-				$secureSettings: ["jwtSecret", "topSecret.password"],
-				jwtSecret: "12345",
-				other: "some",
-				topSecret: {
-					name: "John",
-					password: "password",
-					age: 33
-				}
-			});
-
-			const list = broker.registry.services.list({ onlyLocal: true, skipInternal: true });
-			expect(list[0].settings).toEqual({
-				other: "some",
-				topSecret: {
-					name: "John",
-					age: 33
-				}
-			});
-		});
-	});
-
-	it("should handle undefined settings", () => {
-		delete schema.settings;
-		delete schema.mixins;
-
-		return createService(schema).catch(protectReject).then(svc => {
-			expect(svc.settings).toEqual({});
-
-			const list = broker.registry.services.list({ onlyLocal: true, skipInternal: true });
-			expect(list[0].settings).toEqual({});
-		});
-	});
-
-	it("should handle undefined & defined settings #1", () => {
-		schema.settings = {};
-		delete schema.mixins;
-
-		return createService(schema).catch(protectReject).then(svc => {
-			expect(svc.settings).toEqual({});
-
-			const list = broker.registry.services.list({ onlyLocal: true, skipInternal: true });
-			expect(list[0].settings).toEqual({});
-		});
-	});
-
-	it("should handle undefined & defined settings #1", () => {
-		delete schema.settings;
-		schema.mixins = {};
-
-		return createService(schema).catch(protectReject).then(svc => {
-			expect(svc.settings).toEqual({});
-
-			const list = broker.registry.services.list({ onlyLocal: true, skipInternal: true });
-			expect(list[0].settings).toEqual({});
-		});
-	});
-});
-
-/*
-describe("Test currentContext", () => {
-
-	it("should call broker methods", () => {
-		const broker = new ServiceBroker({ logger: false, internalServices: false });
-		const ctx = new Context(broker);
-		broker.getCurrentContext = jest.fn(() => ctx);
-		broker.setCurrentContext = jest.fn();
-
-		const schema = {
-			name: "posts"
-		};
-
-		const service = broker.createService(schema);
-
-		return service._start()
-			.catch(protectReject)
-			.then(() => {
-				service.currentContext = ctx;
-
-				expect(broker.setCurrentContext).toHaveBeenCalledTimes(1);
-				expect(broker.setCurrentContext).toHaveBeenCalledWith(ctx);
-
-				let res = service.currentContext;
-				expect(res).toBe(ctx);
-				expect(broker.getCurrentContext).toHaveBeenCalledTimes(1);
-				expect(broker.getCurrentContext).toHaveBeenCalledWith();
-			});
-	});
-});
-*/
